requirements.txt
httpx>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
rich>=13.0.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2
aiohttp>=3.9.0
tenacity>=8.2.0
click>=8.1.0
tabulate>=0.9.0
